refactor: migrate to NextAuth v5 beta and centralize authentication logic in auth.ts
