+++ app/(guest-routes)/layout.tsx
@@ -1,14 +1,13 @@
-import { getServerSession } from "next-auth";
+import { auth } from "@/auth";
 import { redirect } from "next/navigation";
 import { ReactNode } from "react";
-import { authOptions } from "../api/auth/[...nextauth]/route";
 
 interface Props {
   children: ReactNode;
 }
 
 export default async function Layout({ children }: Props) {
-  const session = await getServerSession(authOptions);
+  const session = await auth();
   session?.user && redirect("/dashboard");
 
   return <>{children}</>;
