+++ app/(user-routes)/layout.tsx
@@ -1,6 +1,5 @@
-import { getServerSession } from "next-auth";
+import { auth } from "@/auth";
 import { ReactNode } from "react";
-import { authOptions } from "../api/auth/[...nextauth]/route";
 import { redirect } from "next/navigation";
 
 interface Props {
@@ -8,7 +7,7 @@ interface Props {
 }
 
 export default async function Layout({ children }: Props) {
-  const session = await getServerSession(authOptions);
+  const session = await auth();
   !session?.user && redirect("/unauthenticated");
 
   return <div className="pt-8">{children}</div>;
