+++ app/api/records/date/weekly/route.ts
@@ -1,11 +1,10 @@
-import { getServerSession } from "next-auth";
+import { auth } from "@/auth";
 import { NextResponse } from "next/server";
 import prisma from "@/lib/prisma";
-import { authOptions } from "@/app/api/auth/[...nextauth]/route";
 
 export async function GET(_req: Request) {
   try {
-    const session = await getServerSession(authOptions);
+    const session = await auth();
     if (!session?.user?.email)
       return NextResponse.json({ message: "Unauthorized" }, { status: 401 });
