+++ package.json
@@ -13,7 +13,7 @@
   },
   "dependencies": {
     "@hookform/resolvers": "^3.1.1",
-    "@next-auth/prisma-adapter": "^1.0.7",
+    "@auth/prisma-adapter": "^2.7.2",
     "@nivo/line": "^0.87.0",
     "@radix-ui/react-alert-dialog": "^1.0.5",
     "@radix-ui/react-dialog": "^1.0.4",
@@ -40,7 +40,7 @@
     "eslint-config-next": "13.4.8",
     "lucide-react": "^0.258.0",
     "next": "^16.1.6",
-    "next-auth": "^4.22.1",
+    "next-auth": "beta",
     "postcss": "8.4.31",
     "react": "18.2.0",
     "react-day-picker": "^8.8.0",
